[refactor] #56 중복 로직 제거
- SNMRequestType 파일을 분리했습니다.
- 중복되는 로직 제거를 위해 isVaild 속성과 URLRequest 익스텐션을 추가했습니다.
